practice page refactor
