+++ frontend/src/components/practice/static/languageConfiguration.ts
@@ -1,0 +1,20 @@
+export const LANGUAGES = [
+  {
+    id: "javascript",
+    name: "JavaScript",
+    monacoId: "javascript",
+    extension: "js",
+  },
+  { id: "python", name: "Python", monacoId: "python", extension: "py" },
+  { id: "java", name: "Java", monacoId: "java", extension: "java" },
+  { id: "cpp", name: "C++", monacoId: "cpp", extension: "cpp" },
+  { id: "c", name: "C", monacoId: "c", extension: "c" },
+  {
+    id: "typescript",
+    name: "TypeScript",
+    monacoId: "typescript",
+    extension: "ts",
+  },
+  { id: "go", name: "Go", monacoId: "go", extension: "go" },
+  { id: "rust", name: "Rust", monacoId: "rust", extension: "rs" },
+];
