+++ frontend/src/components/practice/types.ts
@@ -1,0 +1,8 @@
+export interface TestCase {
+  id: string;
+  input: string;
+  expectedOutput: string;
+  actualOutput?: string;
+  status?: "passed" | "failed" | "pending";
+  executionTime?: number;
+}
